test(SV-DETECT-001): fuzz tests for paired-end SV detection
Boundary-exploitation fuzz tests for the PEM-signature SV detector
(DetectSVs/ClassifySV/FindDiscordantPairs): identical genomes (all
concordant FR ⇒ no SV called), overlapping vs distant discordant
clusters, empty/null input, span-boundary inclusivity, all five PEM
signatures, and a 3000-trial randomized BE sweep with an independent
spec recomputation of cluster count and support.

The randomized sweep surfaced an overflow bug in
ClusterDiscordantPairs: the cluster-adjacency gap was computed with
Int32 subtraction, so two adjacent discordant pairs with extreme
opposite-sign positions (e.g. a negative coord next to one near
int.MaxValue) overflowed and Math.Abs(int.MinValue) threw
OverflowException. The contract (SV_Detection.md §3.1 positions are
reference coordinates; §3.3/§6.1 only ArgumentNullException is a
documented throw) requires a well-defined result on extreme input.
Widened the subtraction to 64-bit, mirroring the already-fixed
FindBreakpointsIterator.

+++ src/Seqeron/Algorithms/Seqeron.Genomics.Annotation/StructuralVariantAnalyzer.cs
@@ -320,10 +320,14 @@ public static class StructuralVariantAnalyzer
             var prev = pairs[i - 1];
             var curr = pairs[i];
 
+            // The position gaps are computed in 64-bit width: mate positions are
+            // reference coordinates that may span the full Int32 range, and an Int32
+            // subtraction (or Math.Abs of int.MinValue) would overflow and throw on
+            // extreme/opposite-sign coordinates. Mirrors FindBreakpointsIterator.
             bool sameCluster = prev.Chromosome1 == curr.Chromosome1 &&
                                prev.Chromosome2 == curr.Chromosome2 &&
-                               Math.Abs(curr.Position1 - prev.Position1) <= clusterDistance &&
-                               Math.Abs(curr.Position2 - prev.Position2) <= clusterDistance;
+                               Math.Abs((long)curr.Position1 - prev.Position1) <= clusterDistance &&
+                               Math.Abs((long)curr.Position2 - prev.Position2) <= clusterDistance;
 
             if (sameCluster)
             {
